requirements.txt
gspread==5.4.0
oauth2client==4.1.3
python-telegram-bot==20.0
openai==0.27.0
whisper==1.0.0
numpy==1.23.5
torch==1.13.1